+++ CHANGELOG.md
@@ -15,6 +15,10 @@ Changelogs for this project are recorded in this file since v0.2.0.
 
 * Support for Python version 3.8 is dropped
 
+### Added
+
+* `per_timeseries` and `per_feature` options for min-max and mean-variance scalers ([#536](https://github.com/tslearn-team/tslearn/issues/536))
+
 ## [v0.6.4]
 
 ### Changed